refactor(transfer): split disk_commit process into submodules
